// psk-ui/src/app/services/models/daily-calendar-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HourlySessionResponse } from '../models/hourly-session-response';
export interface DailyCalendarResponse {
  dayOfWeek?: 'MONDAY' | 'TUESDAY' | 'WEDNESDAY' | 'THURSDAY' | 'FRIDAY' | 'SATURDAY' | 'SUNDAY';
  sessions?: Array<HourlySessionResponse>;
}